Increased IsGrounded sensitivity
Agent will be considered grounded at a closer distance to the block's surface. Unable to replicate bug where agent is considered on a block when it is touching its side. Next: possibly add whether agent is turning left or right to state inputs because it needs to know when to stop turning. A simpler problem would be to have it always face the nearest block, but that is probably too easy.