Better checks to see if the agent stepped on a block
Uses previous update position to determine if the agent was close to the ground for both updates. If so, we assume that it must have touched the ground and jumped again in between updates.
Added XML observation to only get the block below the agent, and added function to read this observation. Converted current loop checks for getting the block below the player to use the new function.
Added global variable for "Grounded" threshold.
Reduced Tensorflow prints.